Refactor degree 2 extensions to a common module
